+++ src/util.h
@@ -3,15 +3,15 @@
 
 // Print functions
 void print_properties_h(void);
-void print_particle(double *p, double time);
-void print_all_particles(double **P, int numParticle, double time);
+void print_particle(double *p, int time);
+void print_all_particles(double **P, int numParticle, int time);
 
 // Particle generation
 void particles_gen(double **P, int light_cnt, int medium_cnt, int heavy_cnt, int padding_cnt);
 
 // Update particle
-void update(unsigned char* image, double **P, double **P_force, 
-	int total_p_cnt, int img_width, int img_height, 
+void update(unsigned char* image, double **P, double **P_force,
+	int total_p_cnt, int img_width, int img_height,
 	double step_size, int regenerate_img);
